Update header file

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -9,9 +9,9 @@
  *
  * Return: Nothing.
  */
-void array_iterator(int *array, size_t size, void (*action)(int))
+void array_iterator(int *array, int size, void (*action)(int))
 {
-	unsigned int i;
+	int i;
 
 	for (i = 0; i < size; i++)
 		action(array[i]);
